requirements.txt
beautifulsoup4>=4.12
matplotlib>=3.8
playwright>=1.45
PyYAML>=6.0
requests>=2.32
tqdm>=4.66
